MDC-2: refactored success screen to component
