+++ components/thank-you/ThankYou.styled.js
