+++ components/thank-you/index.js
